Separated registration into organization and individual registration pages instead of using the same page
