+++ packages/app/pages/Registration/individualRegistration.tsx
@@ -1,0 +1,18 @@
+/**
+ * @author Connor Hibbs <hibbscm@msoe.edu>
+ * @version 1.0.0
+ */
+
+import * as React from "react";
+import * as EmailValidator from "email-validator";
+
+class IndividualRegistration extends React.Component {
+
+  render() {
+    return (
+      <div>
+        <p>Individual registration coming soon</p>
+      </div>
+    );
+  }
+}
